Refactor flow loading and processing
