+++ src/backend/langflow/__init__.py
@@ -1,9 +1,10 @@
 from importlib import metadata
 
+from langflow.interface.custom.custom_component import CustomComponent
+
 # Deactivate cache manager for now
 # from langflow.services.cache import cache_service
-from langflow.processing.process import load_flow_from_json
-from langflow.interface.custom.custom_component import CustomComponent
+from langflow.processing.load import load_flow_from_json
 
 try:
     __version__ = metadata.version(__package__)
@@ -12,4 +13,4 @@ except metadata.PackageNotFoundError:
     __version__ = ""
 del metadata  # optional, avoids polluting the results of dir(__package__)
 
-__all__ = ["load_flow_from_json", "cache_service", "CustomComponent"]
+__all__ = ["load_flow_from_json", "CustomComponent"]
