+++ tests/test_loading.py
@@ -3,7 +3,7 @@ import json
 import pytest
 from langchain.chains.base import Chain
 from langflow.graph import Graph
-from langflow.processing.process import load_flow_from_json
+from langflow.processing.load import load_flow_from_json
 from langflow.utils.payload import get_root_vertex
